+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Remove deprecated API methods from the last version
 * Finally set a value to the API column
+* No duplicate session login
 
 ######1.9